Add class Region and related

+++ src-java/sondas/src/main/java/com/elo7/probes/domain/Position.java
@@ -1,24 +1,24 @@
 package com.elo7.probes.domain;
 
 public class Position {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
     public Position() {
         this.x = 0;
         this.y = 0;
     }
 
-    public Position(int x, int y) {
+    public Position(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
